+++ src/main/java/tfcmetallum/util/RegistryHandler.java
@@ -103,6 +103,8 @@ public final class RegistryHandler {
                                                                                               // Redstone
     public static final ResourceLocation ENDERIUM = new ResourceLocation(MOD_ID, "enderium"); // Lead + Platinum + Ender
                                                                                               // Pearl
+    public static final ResourceLocation FLUXED_ELECTRUM = new ResourceLocation(MOD_ID, "fluxed_electrum"); // Electrum + Redstone
+
 
     // Ores
     public static final ResourceLocation NATIVE_ARDITE = new ResourceLocation(MOD_ID, "native_ardite");
@@ -320,17 +322,25 @@ public final class RegistryHandler {
             r.register(new Metal(IRIDIUM, Metal.Tier.TIER_III, true, 0.1f, ConfigTFCM.MELT_TEMP.iridium, 0xFFADAFCF,
                     ToolMaterialsTFCM.IRIDIUM, ArmorMaterialsTFCM.IRIDIUM));
         }
+
         if (ConfigTFCM.METALS.lumium) {
             r.register(new Metal(LUMIUM, Metal.Tier.TIER_IV, true, 0.1f, ConfigTFCM.MELT_TEMP.lumium, 0xFFDBDC78,
                     ToolMaterialsTFCM.LUMIUM, ArmorMaterialsTFCM.LUMIUM));
         }
+
         if (ConfigTFCM.METALS.signalum) {
             r.register(new Metal(SIGNALUM, Metal.Tier.TIER_IV, true, 0.1f, ConfigTFCM.MELT_TEMP.signalum, 0xFFF14B00,
                     ToolMaterialsTFCM.SIGNALUM, ArmorMaterialsTFCM.SIGNALUM));
         }
+
         if (ConfigTFCM.METALS.enderium) {
             r.register(new Metal(ENDERIUM, Metal.Tier.TIER_V, true, 0.1f, ConfigTFCM.MELT_TEMP.signalum, 0xFF0A4849,
                     ToolMaterialsTFCM.ENDERIUM, ArmorMaterialsTFCM.ENDERIUM));
+        }
+
+        if (ConfigTFCM.METALS.fluxed_electrum) {
+            r.register(new Metal(FLUXED_ELECTRUM, Metal.Tier.TIER_VI, true, 0.1f, ConfigTFCM.MELT_TEMP.fluxed_electrum, 0xFFC19724,
+                    ToolMaterialsTFCM.FLUXED_ELECTRUM, ArmorMaterialsTFCM.FLUXED_ELECTRUM));
         }
     }